+++ lib/src/ast/sass/supports_condition/anything.dart
@@ -21,5 +21,8 @@ class SupportsAnything implements SupportsCondition {
 
   SupportsAnything(this.contents, this.span);
 
+  Interpolation toInterpolation() =>
+      Interpolation.concat(['(', contents, ')'], span);
+
   String toString() => "($contents)";
 }