+++ lib/src/ast/sass/supports_condition/function.dart
@@ -23,5 +23,8 @@ class SupportsFunction implements SupportsCondition {
 
   SupportsFunction(this.name, this.arguments, this.span);
 
+  Interpolation toInterpolation() =>
+      Interpolation.concat([name, '(', arguments, ')'], span);
+
   String toString() => "$name($arguments)";
 }