+++ lib/src/ast/sass/supports_condition/operation.dart
@@ -5,6 +5,7 @@
 import 'package:meta/meta.dart';
 import 'package:source_span/source_span.dart';
 
+import '../interpolation.dart';
 import '../supports_condition.dart';
 import 'negation.dart';
 
@@ -34,10 +35,25 @@ class SupportsOperation implements SupportsCondition {
     }
   }
 
-  String toString() =>
-      "${_parenthesize(left)} $operator ${_parenthesize(right)}";
+  Interpolation toInterpolation() => Interpolation.concat([
+        ..._parenthesizeInterpolation(left),
+        " $operator ",
+        ..._parenthesizeInterpolation(right)
+      ], span);
 
-  String _parenthesize(SupportsCondition condition) =>
+  /// Returns a list that can be passed to [Interpolation.concat], with
+  /// parentheses around [condition] if necessary.
+  List<Object /* String | Expression | Interpolation */ >
+      _parenthesizeInterpolation(SupportsCondition condition) => condition
+                  is SupportsNegation ||
+              (condition is SupportsOperation && condition.operator == operator)
+          ? ["(", condition.toInterpolation(), ")"]
+          : [condition.toInterpolation()];
+
+  String toString() =>
+      "${_parenthesizeString(left)} $operator ${_parenthesizeString(right)}";
+
+  String _parenthesizeString(SupportsCondition condition) =>
       condition is SupportsNegation ||
               (condition is SupportsOperation && condition.operator == operator)
           ? "($condition)"